Inline headers less than 75 lines
Plan to ideally remove all .i.hh files... but these are the low-hanging
fruit.
