+++ src/comm/detail/MpiTypes.nompi.hh
@@ -3,7 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file MpiTypes.nompi.i.hh
+//! \file MpiTypes.nompi.hh
 //---------------------------------------------------------------------------//
 #pragma once
