--- src/physics/base/CutoffView.i.hh
@@ -1,56 +1,0 @@
-//----------------------------------*-C++-*----------------------------------//
-// Copyright 2021-2022 UT-Battelle, LLC, and other Celeritas developers.
-// See the top-level COPYRIGHT file for details.
-// SPDX-License-Identifier: (Apache-2.0 OR MIT)
-//---------------------------------------------------------------------------//
-//! \file CutoffView.i.hh
-//---------------------------------------------------------------------------//
-
-namespace celeritas
-{
-//---------------------------------------------------------------------------//
-/*!
- * Construct view from host/device for the given material id.
- */
-CELER_FUNCTION
-CutoffView::CutoffView(const CutoffData& params, MaterialId material)
-    : params_(params), material_(material)
-{
-    CELER_EXPECT(params_);
-    CELER_EXPECT(material_ < params_.num_materials);
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Return energy cutoff value.
- */
-CELER_FUNCTION auto CutoffView::energy(ParticleId particle) const -> Energy
-{
-    return this->get(particle).energy;
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Return range cutoff value.
- */
-CELER_FUNCTION real_type CutoffView::range(ParticleId particle) const
-{
-    return this->get(particle).range;
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Get the cutoff for the given particle and material.
- */
-CELER_FUNCTION ParticleCutoff CutoffView::get(ParticleId particle) const
-{
-    CELER_EXPECT(particle < params_.id_to_index.size());
-    CELER_EXPECT(params_.id_to_index[particle] < params_.num_particles);
-    CutoffId id{params_.num_materials * params_.id_to_index[particle]
-                + material_.get()};
-    CELER_ENSURE(id < params_.cutoffs.size());
-    return params_.cutoffs[id];
-}
-
-//---------------------------------------------------------------------------//
-} // namespace celeritas
