+++ src/physics/base/ParticleParams.hh
@@ -110,6 +110,54 @@ class ParticleParams
 };
 
 //---------------------------------------------------------------------------//
-} // namespace celeritas
+// INLINE DEFINITIONS
+//---------------------------------------------------------------------------//
+/*!
+ * Get particle name.
+ */
+const std::string& ParticleParams::id_to_label(ParticleId id) const
+{
+    CELER_EXPECT(id < this->size());
+    return md_[id.get()].first;
+}
 
-#include "ParticleParams.i.hh"
+//---------------------------------------------------------------------------//
+/*!
+ * Get PDG code for a particle ID.
+ */
+PDGNumber ParticleParams::id_to_pdg(ParticleId id) const
+{
+    CELER_EXPECT(id < this->size());
+    return md_[id.get()].second;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Find the ID from a name.
+ */
+ParticleId ParticleParams::find(const std::string& name) const
+{
+    auto iter = name_to_id_.find(name);
+    if (iter == name_to_id_.end())
+    {
+        return ParticleId{};
+    }
+    return iter->second;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Find the ID from a PDG code.
+ */
+ParticleId ParticleParams::find(PDGNumber pdg_code) const
+{
+    auto iter = pdg_to_id_.find(pdg_code);
+    if (iter == pdg_to_id_.end())
+    {
+        return ParticleId{};
+    }
+    return iter->second;
+}
+
+//---------------------------------------------------------------------------//
+} // namespace celeritas
