+++ src/physics/em/detail/RBEnergySampler.hh
@@ -7,6 +7,11 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
+#include <cmath>
+
+#include "base/Algorithms.hh"
+#include "random/distributions/BernoulliDistribution.hh"
+#include "random/distributions/ReciprocalDistribution.hh"
 #include "physics/base/Units.hh"
 #include "physics/base/CutoffView.hh"
 #include "physics/base/ParticleTrackView.hh"
@@ -16,6 +21,7 @@
 
 #include "RelativisticBremData.hh"
 #include "RBDiffXsCalculator.hh"
+#include "PhysicsConstants.hh"
 
 namespace celeritas
 {
@@ -58,7 +64,54 @@ class RBEnergySampler
 };
 
 //---------------------------------------------------------------------------//
+// INLINE DEFINITIONS
+//---------------------------------------------------------------------------//
+/*!
+ * Construct from incident particle and energy.
+ */
+CELER_FUNCTION
+RBEnergySampler::RBEnergySampler(const RelativisticBremNativeRef& shared,
+                                 const ParticleTrackView&         particle,
+                                 const CutoffView&                cutoffs,
+                                 const MaterialView&              material,
+                                 const ElementComponentId&        elcomp_id)
+    : calc_dxsec_(shared, particle, material, elcomp_id)
+{
+    // Min and max kinetic energy limits for sampling the secondary photon
+    real_type gamma_cutoff = value_as<Energy>(cutoffs.energy(shared.ids.gamma));
+    real_type inc_energy   = value_as<Energy>(particle.energy());
+
+    tmin_sq_ = ipow<2>(min(gamma_cutoff, inc_energy));
+    tmax_sq_ = ipow<2>(min(value_as<Energy>(high_energy_limit()), inc_energy));
+
+    CELER_ENSURE(tmax_sq_ >= tmin_sq_);
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Sample the bremsstrahlung photon energy based on G4eBremsstrahlungRelModel
+ * of the Geant4 10.7 release.
+ */
+template<class Engine>
+CELER_FUNCTION auto RBEnergySampler::operator()(Engine& rng) -> Energy
+{
+    real_type density_corr = calc_dxsec_.density_correction();
+    ReciprocalDistribution<real_type> sample_exit_esq(tmin_sq_ + density_corr,
+                                                      tmax_sq_ + density_corr);
+
+    // Sampled energy and corresponding cross section for rejection
+    real_type gamma_energy{0};
+    real_type dsigma{0};
+
+    do
+    {
+        gamma_energy = std::sqrt(sample_exit_esq(rng) - density_corr);
+        dsigma       = calc_dxsec_(Energy{gamma_energy});
+    } while (!BernoulliDistribution(dsigma / calc_dxsec_.maximum_value())(rng));
+
+    return Energy{gamma_energy};
+}
+
+//---------------------------------------------------------------------------//
 } // namespace detail
 } // namespace celeritas
-
-#include "RBEnergySampler.i.hh"
