+++ src/physics/em/detail/SBEnergyDistribution.hh
@@ -7,8 +7,10 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
+#include "base/Algorithms.hh"
 #include "physics/base/Units.hh"
 #include "physics/grid/TwodSubgridCalculator.hh"
+#include "random/distributions/BernoulliDistribution.hh"
 #include "SeltzerBergerData.hh"
 #include "SBEnergyDistHelper.hh"
 
@@ -120,7 +122,49 @@ class SBEnergyDistribution
 };
 
 //---------------------------------------------------------------------------//
+// INLINE DEFINITIONS
+//---------------------------------------------------------------------------//
+/*!
+ * Construct from incident particle and energy.
+ *
+ * The incident energy *must* be within the bounds of the SB table data, so the
+ * Model's applicability must be consistent with the table data.
+ */
+template<class X>
+CELER_FUNCTION
+SBEnergyDistribution<X>::SBEnergyDistribution(const SBEnergyDistHelper& helper,
+                                              X scale_xs)
+    : helper_(helper)
+    , inv_max_xs_{1
+                  / (helper.max_xs().value() * scale_xs(helper.max_xs_energy()))}
+    , scale_xs_(::celeritas::move(scale_xs))
+{
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Sample the exiting energy by doing a table lookup and rejection.
+ */
+template<class X>
+template<class Engine>
+CELER_FUNCTION auto SBEnergyDistribution<X>::operator()(Engine& rng) -> Energy
+{
+    // Sampled energy
+    Energy exit_energy;
+    // Calculated cross section used inside rejection sampling
+    real_type xs{};
+    do
+    {
+        // Sample scaled energy and subtract correction factor
+        exit_energy = helper_.sample_exit_energy(rng);
+
+        // Interpolate the differential cross setion at the sampled exit energy
+        xs = helper_.calc_xs(exit_energy).value() * scale_xs_(exit_energy);
+        CELER_ASSERT(xs >= 0 && xs <= 1 / inv_max_xs_);
+    } while (!BernoulliDistribution(xs * inv_max_xs_)(rng));
+    return exit_energy;
+}
+
+//---------------------------------------------------------------------------//
 } // namespace detail
 } // namespace celeritas
-
-#include "SBEnergyDistribution.i.hh"
