+++ src/physics/grid/UniformGrid.hh
@@ -7,6 +7,7 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
+#include "base/Assert.hh"
 #include "base/Macros.hh"
 #include "base/Types.hh"
 #include "UniformGridData.hh"
@@ -56,6 +57,43 @@ class UniformGrid
 };
 
 //---------------------------------------------------------------------------//
-} // namespace celeritas
+// INLINE DEFINITIONS
+//---------------------------------------------------------------------------//
+/*!
+ * Construct with data.
+ */
+CELER_FUNCTION
+UniformGrid::UniformGrid(const UniformGridData& data) : data_(data)
+{
+    CELER_EXPECT(data_);
+}
 
-#include "UniformGrid.i.hh"
+//---------------------------------------------------------------------------//
+/*!
+ * Get the value at the given grid point.
+ */
+CELER_FUNCTION auto UniformGrid::operator[](size_type i) const -> value_type
+{
+    CELER_EXPECT(i < data_.size);
+    return data_.front + data_.delta * i;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Find the value bin such that data[result] <= value < data[result + 1].
+ *
+ * The given value *must* be in range, because out-of-bounds values usually
+ * require different treatment (e.g. clipping to the boundary values rather
+ * than interpolating). It's easier to test the exceptional cases (final grid
+ * point) outside of the grid view.
+ */
+CELER_FUNCTION size_type UniformGrid::find(value_type value) const
+{
+    CELER_EXPECT(value >= this->front() && value < this->back());
+    auto bin = static_cast<size_type>((value - data_.front) / data_.delta);
+    CELER_ENSURE(bin + 1 < this->size());
+    return bin;
+}
+
+//---------------------------------------------------------------------------//
+} // namespace celeritas
